Adjusted report death 1st page and unit tests
