+++ src/applications/disability-benefits/686c-674/components/DependentViewField.jsx
@@ -4,7 +4,7 @@ export default function DependentField({ formData }) {
   const { first, middle, last, suffix } = formData.fullName;
 
   return (
-    <div>
+    <div className="vads-u-padding--2">
       <strong>
         {first} {middle && `${middle} `}
         {last}
