+++ src/applications/disability-benefits/686c-674/config/chapters/report-dependent-death/dependent-additional-information/dependent-additional-information.js
@@ -13,31 +13,6 @@ export const schema = {
       items: {
         type: 'object',
         properties: {
-          dependentType: {
-            type: 'string',
-            enum: ['SPOUSE', 'DEPENDENT_PARENT', 'CHILD'],
-            enumNames: ['Spouse', 'Dependent Parent', 'Child'],
-          },
-          childStatus: {
-            type: 'object',
-            properties: {
-              childUnder18: {
-                type: 'boolean',
-              },
-              stepChild: {
-                type: 'boolean',
-              },
-              adopted: {
-                type: 'boolean',
-              },
-              disabled: {
-                type: 'boolean',
-              },
-              childOver18InSchool: {
-                type: 'boolean',
-              },
-            },
-          },
           deceasedDateOfDeath: genericSchemas.date,
           deceasedLocationOfDeath: genericSchemas.genericLocation,
         },
@@ -46,38 +21,28 @@ export const schema = {
   },
 };
 
+/*
+        updateUiSchema: formData => {
+          console.log(formData)
+          if (formData['view:selectable686Options']['reportDeath'] === true) {
+            return {
+              'ui:validations': validateBooleanGroup,
+              'ui:errorMessages': {
+                atLeastOne: 'You must have expenses for at least one benefit.',
+              },
+              'ui:required': () => true,
+              'ui:title': "Child's status (Check all that apply) *(required)",
+            };
+          } else {
+            
+          }
+        },
+*/
+
 export const uiSchema = {
   deaths: {
     'ui:options': { viewField: DependentViewField },
     items: {
-      dependentType: {
-        'ui:title': "What was your dependent's status?",
-        'ui:widget': 'radio',
-      },
-      childStatus: {
-        'ui:title': "Child's status (Check all that apply)",
-        'ui:options': {
-          expandUnder: 'dependentType',
-          expandUnderCondition: 'CHILD',
-          showFieldLabel: true,
-          keepInPageOnReview: true,
-        },
-        childUnder18: {
-          'ui:title': 'Child under 18',
-        },
-        stepChild: {
-          'ui:title': 'Stepchild',
-        },
-        adopted: {
-          'ui:title': 'Adopted',
-        },
-        disabled: {
-          'ui:title': 'Disabled',
-        },
-        childOver18InSchool: {
-          'ui:title': 'Child over 18 in school',
-        },
-      },
       deceasedDateOfDeath: currentOrPastDateUI('Dependent’s date of death'),
       deceasedLocationOfDeath: {
         'ui:title': 'Place of death',
