+++ src/applications/disability-benefits/686c-674/config/chapters/report-dependent-death/dependent-information/helpers.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export function childInfo({ formData }) {
+  return (
+    <div className="vads-u-padding--2">
+      <strong>
+        {formData.first} {formData.last}
+      </strong>
+      <br />
+    </div>
+  );
+}
