+++ src/applications/disability-benefits/686c-674/tests/config/reportDeath.unit.spec.jsx
@@ -21,7 +21,7 @@ describe('686 report dependent death', () => {
     'view:selectable686Options': {
       reportDeath: true,
     },
-  }
+  };
 
   it('should render', () => {
     const form = mount(
@@ -32,7 +32,7 @@ describe('686 report dependent death', () => {
         data={formData}
       />,
     );
-    expect(form.find('input').length).to.equal(3);
+    expect(form.find('input').length).to.equal(6);
     expect(form.find('select').length).to.equal(1);
     form.unmount();
   });
@@ -49,9 +49,115 @@ describe('686 report dependent death', () => {
       />,
     );
     form.find('form').simulate('submit');
-    expect(form.find('.usa-input-error').length).to.equal(2);
+    expect(form.find('.usa-input-error').length).to.equal(3);
     expect(onSubmit.called).to.be.false;
     form.unmount();
   });
 
+  it('select spouse as dependentType', () => {
+    const form = mount(
+      <DefinitionTester
+        schema={schema}
+        definitions={formConfig.defaultDefinitions}
+        uiSchema={uiSchema}
+        data={formData}
+      />,
+    );
+    selectRadio(form, 'root_deaths_0_dependentType', 'SPOUSE');
+    expect(form.find('input').length).to.equal(6);
+    form.unmount();
+  });
+
+  it('select dependent parent as dependentType', () => {
+    const form = mount(
+      <DefinitionTester
+        schema={schema}
+        definitions={formConfig.defaultDefinitions}
+        uiSchema={uiSchema}
+        data={formData}
+      />,
+    );
+    selectRadio(form, 'root_deaths_0_dependentType', 'DEPENDENT_PARENT');
+    expect(form.find('input').length).to.equal(6);
+    form.unmount();
+  });
+
+  it('should expand child options if dependentType is child', () => {
+    const form = mount(
+      <DefinitionTester
+        schema={schema}
+        definitions={formConfig.defaultDefinitions}
+        uiSchema={uiSchema}
+        data={formData}
+      />,
+    );
+    selectRadio(form, 'root_deaths_0_dependentType', 'CHILD');
+    expect(form.find('input').length).to.equal(11);
+    form.unmount();
+  });
+
+  it('should submit a valid form with a dependent spouse', () => {
+    const onSubmit = sinon.spy();
+    const form = mount(
+      <DefinitionTester
+        schema={schema}
+        definitions={formConfig.defaultDefinitions}
+        uiSchema={uiSchema}
+        onSubmit={onSubmit}
+        data={formData}
+      />,
+    );
+    // dependent type
+    selectRadio(form, 'root_deaths_0_dependentType', 'SPOUSE');
+    fillData(form, 'input#root_deaths_0_fullName_first', 'Billy');
+    fillData(form, 'input#root_deaths_0_fullName_last', 'Bob');
+
+    form.find('form').simulate('submit');
+    expect(form.find('.usa-input-error').length).to.equal(0);
+    expect(onSubmit.called).to.be.true;
+    form.unmount();
+  });
+
+  it('should submit a valid form with a dependent child', () => {
+    const onSubmit = sinon.spy();
+    const form = mount(
+      <DefinitionTester
+        schema={schema}
+        definitions={formConfig.defaultDefinitions}
+        uiSchema={uiSchema}
+        onSubmit={onSubmit}
+        data={formData}
+      />,
+    );
+    // dependent type
+    selectRadio(form, 'root_deaths_0_dependentType', 'CHILD');
+    selectCheckbox(form, 'root_deaths_0_childStatus_childUnder18', true);
+    fillData(form, 'input#root_deaths_0_fullName_first', 'Billy');
+    fillData(form, 'input#root_deaths_0_fullName_last', 'Bob');
+
+    form.find('form').simulate('submit');
+    expect(form.find('.usa-input-error').length).to.equal(0);
+    expect(onSubmit.called).to.be.true;
+    form.unmount();
+  });
+
+  it('should not submit when child is selected without any subtypes', () => {
+    const onSubmit = sinon.spy();
+    const form = mount(
+      <DefinitionTester
+        schema={schema}
+        definitions={formConfig.defaultDefinitions}
+        uiSchema={uiSchema}
+        onSubmit={onSubmit}
+        data={formData}
+      />,
+    );
+    // dependent type
+    selectRadio(form, 'root_deaths_0_dependentType', 'CHILD');
+
+    form.find('form').simulate('submit');
+    expect(form.find('.usa-input-error').length).to.equal(3);
+    expect(onSubmit.called).to.be.false;
+    form.unmount();
+  });
 });
